DBZ-1478 Avoiding split package between runtime and deployment modules
